fix resize mode issue
