+++ library/build.gradle
@@ -13,8 +13,8 @@ android {
     defaultConfig {
         minSdkVersion 16
         targetSdkVersion 29
-        versionCode 18
-        versionName "2.0.2"
+        versionCode 19
+        versionName "2.0.3"
         vectorDrawables.useSupportLibrary = true
         multiDexEnabled true
